src/requirements.txt
cloudshell-shell-core>=2.3.0,<2.4.0
cloudshell-pdu-core>=1.0.0,<2.0.0
cloudshell-snmp==3.1.2
cloudshell-automation-api>=8.1.0,<8.2.0